% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{cross}
\alias{cross}
\title{Cross Product of two}
\usage{
cross(a, b)
}
\arguments{
\item{a}{Length-3 numeric vector.}

\item{b}{Length-3 numeric vector.}
}
\description{
Cross Product of two
}
\keyword{internal}
