% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_example_molecule.R
\name{get_example_molecule}
\alias{get_example_molecule}
\title{Get Example Molecule}
\usage{
get_example_molecule(molecule)
}
\arguments{
\item{molecule}{One of the built-in SDF files. These are "benzene", "buckyball",
"caffeine", "capsaicin", "cinnemaldehyde", "geraniol", "luciferin", "morphine",
"penicillin", "pfoa", "skatole", "tubocurarine_chloride".}
}
\value{
List giving the atom locations and the connections between atoms.
}
\description{
Loads the structure of the built-in molecules. All SDF files obtained from Pubchem.
txt extension only included to pass R CHECK.
}
\examples{
get_example_molecule("benzene")
get_example_molecule("cinnemaldehyde")
get_example_molecule("geraniol")
}
