% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sdf.R
\name{read_sdf}
\alias{read_sdf}
\title{Read SDF File}
\usage{
read_sdf(filename)
}
\arguments{
\item{filename}{Filename to the sdf file.}
}
\value{
List giving the atom locations and the connections between atoms.
}
\description{
Reads an SDF file and extracts the 3D molecule model
}
\examples{
#This assumes a hypothetical SDF file in your working directory:
if(file.exists("molecule.sdf")) {
  read_pdb("molecule.sdf") \%>\%
    generate_full_scene() \%>\%
    render_model()
}
}
