% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_atom_scene.R
\name{generate_atom_scene}
\alias{generate_atom_scene}
\title{Build Rayrender Scene (atoms only)}
\usage{
generate_atom_scene(model, scale = 1, center = TRUE)
}
\arguments{
\item{model}{Model extracted from a PDB or SDF file.}

\item{scale}{Default `1`. Amount to scale the inter-atom spacing.}

\item{center}{Default `TRUE`. Centers the bounding box of the model.}
}
\value{
Rayrender scene (tibble) containing only the atoms in a molecule/protein.
}
\description{
Reads an SDF file and extracts the 3D molecule model
}
\examples{
#Generate a scene with caffeine molecule with just the atoms
\donttest{
get_example_molecule("caffeine") \%>\%
  read_sdf() \%>\%
  generate_atom_scene() \%>\%
  render_model()

#Generate a scene with caffeine, reducing the inter-atom spacing
get_example_molecule("caffeine") \%>\%
  read_sdf() \%>\%
  generate_atom_scene(scale=0.5) \%>\%
  render_model()
}
}
