% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_full_scene.R
\name{generate_full_scene}
\alias{generate_full_scene}
\title{Build Rayrender Scene (bonds + atoms)}
\usage{
generate_full_scene(
  model,
  x = 0,
  y = 0,
  z = 0,
  scale = 1,
  center = TRUE,
  force_single_bonds = FALSE
)
}
\arguments{
\item{model}{Model extracted from a PDB or SDF file.}

\item{x}{Default `0`. X offset, applied after centering.}

\item{y}{Default `0`. Y offset, applied after centering.}

\item{z}{Default `0`. Z offset, applied after centering.}

\item{scale}{Default `1`. Amount to scale the interatom spacing.}

\item{center}{Default `TRUE`. Centers the bounding box of the model.}

\item{force_single_bonds}{Default `FALSE`. Whether to force all bonds to show as a single connection.}
}
\value{
Rayrender scene (tibble) containing the atoms and bonds in a molecule/protein.
}
\description{
Reads an SDF file and extracts the 3D molecule model
}
\examples{
# Generate a scene with caffeine molecule
\donttest{
get_example_molecule("caffeine") \%>\%
  read_sdf() \%>\%
  generate_full_scene() \%>\%
  render_model()

# Generate a scene with morphine, increasing the inter-atom spacing
get_example_molecule("tubocurarine_chloride") \%>\%
  read_sdf() \%>\%
  generate_full_scene(scale=1.5) \%>\%
  render_model()

# Force bonds to appear as a single link (to focus purely on the shape of the molecule)
get_example_molecule("tubocurarine_chloride") \%>\%
  read_sdf() \%>\%
  generate_full_scene(force_single_bonds = TRUE) \%>\%
  render_model()
}
}
